move the non-vic3 parts of AdvancedEntity into a general IconEntity
